// src/shared/api/orval/types/region.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Smarter Store API
 * Smarter Store 백엔드 API 문서

**관리자 대시보드**: [매출 현황 대시보드](/admin/dashboard.html)
 * OpenAPI spec version: 1.0.0
 */

/**
 * 대한민국 17개 광역자치단체 (행정구역)
 */
export enum Region {
  SEOUL= 'SEOUL',
  INCHEON= 'INCHEON',
  DAEJEON= 'DAEJEON',
  DAEGU= 'DAEGU',
  GWANGJU= 'GWANGJU',
  ULSAN= 'ULSAN',
  BUSAN= 'BUSAN',
  SEJONG= 'SEJONG',
  GYEONGGI= 'GYEONGGI',
  GANGWON= 'GANGWON',
  CHUNGBUK= 'CHUNGBUK',
  CHUNGNAM= 'CHUNGNAM',
  JEONBUK= 'JEONBUK',
  JEONNAM= 'JEONNAM',
  GYEONGBUK= 'GYEONGBUK',
  GYEONGNAM= 'GYEONGNAM',
  JEJU= 'JEJU',

}